make custom shortcuts mostly work, some fixes still needed

+++ Utils.m
@@ -2,7 +2,7 @@
 
 @implementation NSString (FlickPlus)
 - (NSString *)sliceAfterLastUnderscore {
-	int lastUnderscore = [self rangeOfString:@"_" options:NSBackwardsSearch].location;
+	NSUInteger lastUnderscore = [self rangeOfString:@"_" options:NSBackwardsSearch].location;
 	if (lastUnderscore != NSNotFound)
 		return [self substringFromIndex:lastUnderscore + 1];
 	else
